feat: add connection to push transaction in db

+++ script.sql
@@ -1,9 +1,8 @@
 CREATE TABLE clientes (
     id SERIAL PRIMARY KEY,
-    limit_value INT NOT NULL,
-    current INT NOT NULL
+    limit_value INT8 NOT NULL,
+    current INT8 NOT NULL
 );
-
 
 DO $$ BEGIN
 INSERT INTO
@@ -14,13 +13,11 @@ VALUES
     (10000 * 100, 0),
     (100000 * 100, 0),
     (5000 * 100, 0);
-
 END $$;
-
 
 CREATE TABLE transactions (
     client_id INT REFERENCES clientes(id),
-    value INT NOT NULL,
+    value INT8 NOT NULL,
     type CHAR NOT NULL,
     description VARCHAR(10) NOT NULL,
     timestamp TIMESTAMP NOT NULL
